Unit tests: extended unit tests to cover multifield ext funcs
Added several tests to test for functions returning a multifield value,
getting passed a multifield value, or both at the same time. This change
requires the introduction of a CLIPS file used during unit tests because
a multifield value passed into an external function can not be
constructed as part of the extern function call. Hence we call a mediate
function which constructs the multifield value and calls the external
function passing this value.