Environment: fix bug regarding function restriction string
The function string was created in stack memory. Consecutive calls to
add_function() would re-use the very same string, causing previously
added functions of the same arity to have a wrong signature assigned.
Additionally strings were not properly null-terminated and the array was
short by one for the 7 argument form.