fact: add comparison operators and refcount accessor
Added == operator for facts (two facts are the same if their index is
the same), and an accessor function for the reference count of a fact
(this can be used to query if a fact is no longer used within CLIPS,
i.e. if the refcount is 1 then the Fact instance is the only user of the
fact).